requirements.txt
daphne==4.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crcmod==1.7
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django5-aliyun-oss==1.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jmespath==0.10.0
kombu==5.5.4
msgpack==1.1.2
oauthlib==3.3.1
oss2==2.19.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
soupsieve==2.8
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xpinyin==0.7.7
zope.interface==8.1.1
